Fedora

+++ install.sh
@@ -14,13 +14,26 @@ LSB_DIST="$(echo "$LSB_DIST" | tr '[:upper:]' '[:lower:]')"
 
 case "$LSB_DIST" in
     centos|fedora)
-        su -c "
+        SELINUXSTATUS=$(getenforce)
+        if [ "$SELINUXSTATUS" == "Permissive" ] || [ "$SELINUXSTATUS" == "Enforcing" ]; then
+            setsebool -P httpd_can_network_connect 1
+        fi;
+        dnf -y update
+        dnf -y install cockpit-docker docker-compose
+        systemctl start docker
+        systemctl enable docker
+        if systemctl is-active httpd; then
             systemctl stop httpd
+        fi
+        if systemctl is-enabled httpd; then
             systemctl disable httpd
-            dnf -y install nginx
-            systemctl start nginx
-            systemctl enable nginx
-        "
+        fi
+        dnf -y install nginx
+        systemctl start nginx
+        systemctl enable nginx
+        firewall-cmd --permanent --add-service=https
+        firewall-cmd --permanent --add-service=http
+        firewall-cmd --reload
     ;;
     rhel|ol|sles)
         echo "Operating system not supported."
